test[api]: shard the CosyVoice3 enhancer tests onto the cosyvoice3 shard
The model-backed CosyVoice3 enhancer tests cannot live in
lavasr-enhancer.test.js: that file now runs on the 'lavasr' Device Farm
shard, which stages Chatterbox, Supertonic and the enhancer GGUF but never
CosyVoice3's ~2.4 GB model dir. Move them to their own file listed on the
'cosyvoice3' shard, and map the runner to the CosyVoice3 models plus the
enhancer so its shard pre-stages both.
