+++ packages/tts-ggml/test/integration/lavasr-enhancer.test.js
@@ -8,9 +8,11 @@
 // chunk seam-free), the denoiser stays batch-only, and a misconfigured enhancer
 // can't silently become a no-op (an unknown enhancer.type throws). The
 // model-backed tests assert the enhanced output is reported as 48 kHz for
-// Supertonic, Chatterbox, Parler and CosyVoice3, covering the native chunk
-// streaming of the engines that support it; they are gated on the converted
-// enhancer GGUF being staged, and skip cleanly otherwise.
+// Supertonic, Chatterbox and Parler, covering the native chunk streaming of the
+// latter two; they are gated on the converted enhancer GGUF being staged, and
+// skip cleanly otherwise. CosyVoice3's model-backed pair lives in
+// cosyvoice3-lavasr-enhancer.test.js, which the mobile suite shards onto the
+// only row that stages the CosyVoice3 model.
 //
 // Stage the enhancer GGUF via scripts/convert-lavasr-enhancer-to-gguf.py (from
 // the public LavaSRcpp ONNX release) into models/lavasr/lavasr-enhancer.gguf,
@@ -27,16 +29,12 @@ const {
   ensureLavaSRDenoiserGguf,
   ensureSupertonicModel,
   ensureChatterboxModels,
-  ensureParlerModel,
-  ensureCosyvoiceModel
+  ensureParlerModel
 } = require('../utils/downloadModel')
 const { resolveRefWavPath } = require('../utils/runChatterboxTTS')
 
 const platform = os.platform()
 const isMobile = platform === 'ios' || platform === 'android'
-// CosyVoice3 segfaults at load/synthesis on the win32-x64 desktop lane; mirrors
-// the skip in cosyvoice3.test.js so the Windows lane stays green.
-const SKIP_COSYVOICE = platform === 'win32'
 // Mirrors gpu-smoke.test.js: CI runners without a real GPU export NO_GPU=true to
 // skip the GPU-gated entries; QVAC_TTS_GPU_SMOKE_RELAX=1 downgrades a GPU->CPU
 // fallback from a failure to a warning (e.g. a Linux host with no Vulkan SDK).
@@ -895,115 +893,6 @@ test(
       0,
       'useGPU:false -> enhancer on CPU (enhancerBackendDevice=0)'
     )
-  }
-)
-
-// CosyVoice3 is CPU-only in this iteration, and the addon hands the enhancer the
-// engine's *resolved* device, so these also pin that the enhancer really loaded
-// and ran (enhancerBackendDevice=0) rather than silently staying off. Text is
-// kept short to bound CPU LM-decode time in CI, as in cosyvoice3.test.js.
-
-test(
-  'CosyVoice3 + LavaSR enhancer (batch) reports 48 kHz enhanced output',
-  { timeout: 900000, skip: SKIP_COSYVOICE },
-  async (t) => {
-    const baseDir = getBaseDir()
-    const enh = await ensureLavaSREnhancerGguf({
-      targetDir: path.join(baseDir, 'models', 'lavasr')
-    })
-    if (!enh.success) {
-      t.comment('LavaSR enhancer GGUF not staged; skipping.')
-      t.pass('skipped — no enhancer GGUF')
-      return
-    }
-    const dl = await ensureCosyvoiceModel({
-      targetDir: path.join(baseDir, 'models', 'cosyvoice3')
-    })
-    if (!dl.success) {
-      t.fail('CosyVoice3 model files not available — registry fetch failed.')
-      return
-    }
-
-    const model = new TTSGgml({
-      engine: TTSGgml.ENGINE_COSYVOICE3,
-      files: { cosyvoiceModelDir: dl.modelDir, lavasrEnhancer: enh.path },
-      config: { language: 'en', useGPU: false },
-      opts: { stats: true }
-    })
-    await model.load()
-    try {
-      const r = await runAndCollect(model, 'Hello from CosyVoice.')
-      t.is(r.sampleRate, 48000, 'enhanced cosyvoice3 output reports 48 kHz (native is 24 kHz)')
-      t.ok(r.samples > 0, 'enhanced synthesis produced audio')
-      t.ok(r.stats, 'runtimeStats returned (constructed with stats:true)')
-      t.is(
-        r.stats.enhancerBackendDevice,
-        0,
-        'enhancer loaded and ran on CPU (enhancerBackendDevice=0, not -1 = never loaded)'
-      )
-    } finally {
-      try {
-        await model.unload()
-      } catch (_e) {}
-    }
-  }
-)
-
-test(
-  'CosyVoice3 + LavaSR enhancer + native chunk streaming emits 48 kHz chunks',
-  { timeout: 900000, skip: SKIP_COSYVOICE },
-  async (t) => {
-    const baseDir = getBaseDir()
-    const enh = await ensureLavaSREnhancerGguf({
-      targetDir: path.join(baseDir, 'models', 'lavasr')
-    })
-    if (!enh.success) {
-      t.comment('LavaSR enhancer GGUF not staged; skipping.')
-      t.pass('skipped — no enhancer GGUF')
-      return
-    }
-    const dl = await ensureCosyvoiceModel({
-      targetDir: path.join(baseDir, 'models', 'cosyvoice3')
-    })
-    if (!dl.success) {
-      t.fail('CosyVoice3 model files not available — registry fetch failed.')
-      return
-    }
-
-    const model = new TTSGgml({
-      engine: TTSGgml.ENGINE_COSYVOICE3,
-      files: { cosyvoiceModelDir: dl.modelDir, lavasrEnhancer: enh.path },
-      streamChunkTokens: 25, // native chunk streaming + enhancer (the path)
-      config: { language: 'en', useGPU: false },
-      opts: { stats: true }
-    })
-    await model.load()
-    try {
-      const updates = []
-      const response = await model.run({ input: 'Hello from CosyVoice.', type: 'text' })
-      await response
-        .onUpdate((d) => {
-          if (d && d.outputArray) updates.push(d)
-        })
-        .await()
-
-      const total = updates.reduce((acc, u) => acc + u.outputArray.length, 0)
-      t.ok(updates.length >= 1, 'streamed at least one chunk event')
-      t.ok(total > 0, 'streamed enhanced audio produced samples')
-      // Every chunk carrying audio must be tagged at the enhanced 48 kHz rate
-      // rather than CosyVoice3's native 24 kHz — the mislabel this path prevents.
-      for (const u of updates) {
-        if (u.outputArray.length > 0 && u.sampleRate != null) {
-          t.is(u.sampleRate, 48000, 'streamed enhanced chunk reports 48 kHz')
-        }
-      }
-      const isLastCount = updates.filter((u) => u.isLast === true).length
-      t.ok(isLastCount <= 1, 'at most one isLast=true across streamed chunks')
-    } finally {
-      try {
-        await model.unload()
-      } catch (_e) {}
-    }
   }
 )
