+++ packages/tts-ggml/test/mobile/integration.auto.cjs
@@ -22,6 +22,10 @@ async function runChatterboxSpeedTest (options = {}) { // eslint-disable-line no
   return runIntegrationModule('../integration/chatterbox-speed.test.js', options)
 }
 
+async function runCosyvoice3LavasrEnhancerTest (options = {}) { // eslint-disable-line no-unused-vars
+  return runIntegrationModule('../integration/cosyvoice3-lavasr-enhancer.test.js', options)
+}
+
 async function runCosyvoice3Test (options = {}) { // eslint-disable-line no-unused-vars
   return runIntegrationModule('../integration/cosyvoice3.test.js', options)
 }
@@ -75,6 +79,7 @@ module.exports = {
   runChatterboxKvCacheGpuTest,
   runChatterboxMtlTest,
   runChatterboxSpeedTest,
+  runCosyvoice3LavasrEnhancerTest,
   runCosyvoice3Test,
   runGpuSmokeTest,
   runLavasrEnhancerTest,
