+++ packages/tts-ggml/test/mobile/test-groups.json
@@ -25,6 +25,7 @@
       "runLavasrEnhancerTest"
     ],
     "cosyvoice3": [
+      "runCosyvoice3LavasrEnhancerTest",
       "runCosyvoice3Test"
     ],
     "cross-model-compatibility": [
@@ -58,6 +59,7 @@
       "runLavasrEnhancerTest"
     ],
     "cosyvoice3": [
+      "runCosyvoice3LavasrEnhancerTest",
       "runCosyvoice3Test"
     ],
     "cross-model-compatibility": [
